requirements.txt
python>=3.9
numpy
pandas
netcdf4
xarray
python-dateutil
rch
